Stripe

+++ products/models.py
@@ -6,7 +6,7 @@ class Item(models.Model):
     objects = None
     name = models.CharField(max_length=100)
     description = models.TextField()
-    price = models.IntegerField(default=0)
+    price = models.DecimalField(max_digits=10, decimal_places=2, default=0.0)
 
     def __str__(self):
         return self.name
